misc: Refactor (but not working properly due to other conflicts)
